+++ docs/release_notes.rst
@@ -8,6 +8,8 @@ Release Notes
 * Added ``JNEventLoop``
 * Added module ``timers``
 * Added function ``timers.clearImmediate(immediate)``
+* Added function ``timers.clearInterval(timeout)``
+* Added function ``timers.clearTimeout(timeout)``
 * Added function ``timers.setImmediate(callback[, ...args])``
 * Added function ``timers.setInterval(callback[, delay[, ...args]])``
 * Added function ``timers.setTimeout(callback[, delay[, ...args]])``